+++ library/src/components-private/schema/attributes/AttributeProperties.vue
@@ -23,7 +23,7 @@
                   .nio-slat-title-subtitle
                     .nio-slat-title.nio-h4.text-primary-darker.nio-bold(:style="{width: slatWidth}") {{ propertyName }}
                     .nio-slat-subtitle.nio-p.text-primary-dark(v-if="properties[propertyName].description" :style="{width: slatWidth}") {{ properties[propertyName].description }}
-                .property-settings(v-if="displayOnly && properties[propertyName].type !== 'object' && properties[propertyName].type !=='array'")
+                .property-settings(v-if="!hideIndicators && displayOnly && properties[propertyName].type !== 'object' && properties[propertyName].type !=='array'")
                   .pills-container
                     NioPill(
                       list-item
@@ -39,7 +39,7 @@
                       :iconColor="properties[propertyName].filterable ? '#43B463' : '#AEB9E8'"
                       :class="properties[propertyName].filterable ? '' : 'negative'"
                     ) Filterable
-                .property-actions(v-else-if="!displayOnly")
+                .property-actions(v-else-if="!hideIndicators && !displayOnly")
                   NioSwitch(
                     @click.stop=""
                     v-model="properties[propertyName].deliverable"
@@ -96,7 +96,8 @@ export default {
   props: {
     "properties": { type: Object, required: false, default: null},
     "display-only": { type: Boolean, required: false, default: false },
-    "nest": { type: Number, required: true}
+    "nest": { type: Number, required: true},
+    "hideIndicators": { type: Boolean, required: false, default: false }
   },
   data: () => ({
     openPanels: []