+++ CHANGELOG.md
@@ -100,3 +100,4 @@ Added default slot to NioAlert
 v0.4.0
 Added **NioDestinationConnectorControl** for content common to all destination connectors
 Added generic **NioConnectorSettingsControl** for use in all connectors
+Added **utility-ellipsis-vertical** icon