+++ CHANGELOG.md
@@ -101,3 +101,4 @@ v0.4.0
 Added **NioDestinationConnectorControl** for content common to all destination connectors
 Added generic **NioConnectorSettingsControl** for use in all connectors
 Added **utility-ellipsis-vertical** icon
+Added **NioAttributesList** and **NioAttributeProperties** to private components