+++ CHANGELOG.md
@@ -105,3 +105,4 @@ Added **NioAttributesList** and **NioAttributeProperties** to private components
 Added **NioSchema** and **NioSchemaProperties** components
 Fixed bug: set text input to null on model update in **NioAutocomplete**
 Fixed bug: clicking append icon in **NioDatepicker** will launch datepicker
+Added model binding for name and description **NioDestinationConnectorSettings**, updated copy for descriptions